Update to latest crow, resolve link errors.
This resolves compatibility issues on the (as of writing) latest version of crow, and resolves linker errors related to the kTypeNames and kSelf variables.

This also resolves the TODO of disabling the 'Server' header.
